feat: product service merge

+++ product/src/main/java/store/_0982/product/application/dto/ProductRegisterInfo.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 public record  ProductRegisterInfo(
         UUID productId,
         String name,
-        int price,
+        Long price,
         ProductCategory category,
         String description,
         int stock,
